tipo double

+++ src/main/java/app/salesorderdetall/domain/entity/Saleorderdetall.java
@@ -6,7 +6,7 @@ public class Saleorderdetall {
 	private long sale_order_id; 
 	private long product_id; 
 	private int quantity;
-	private int price;
+	private double price;
 	private String currency; 
 	private int status;
 	
@@ -17,9 +17,11 @@ public Saleorderdetall() {
 }
 
 
+
 public long getId() {
 	return id;
 }
+
 
 
 public void setId(long id) {
@@ -27,9 +29,11 @@ public void setId(long id) {
 }
 
 
+
 public long getSale_order_id() {
 	return sale_order_id;
 }
+
 
 
 public void setSale_order_id(long sale_order_id) {
@@ -37,9 +41,11 @@ public void setSale_order_id(long sale_order_id) {
 }
 
 
+
 public long getProduct_id() {
 	return product_id;
 }
+
 
 
 public void setProduct_id(long product_id) {
@@ -47,9 +53,11 @@ public void setProduct_id(long product_id) {
 }
 
 
+
 public int getQuantity() {
 	return quantity;
 }
+
 
 
 public void setQuantity(int quantity) {
@@ -57,14 +65,17 @@ public void setQuantity(int quantity) {
 }
 
 
-public int getPrice() {
+
+public double getPrice() {
 	return price;
 }
 
 
-public void setPrice(int price) {
+
+public void setPrice(double price) {
 	this.price = price;
 }
+
 
 
 public String getCurrency() {
@@ -72,9 +83,11 @@ public String getCurrency() {
 }
 
 
+
 public void setCurrency(String currency) {
 	this.currency = currency;
 }
+
 
 
 public int getStatus() {
@@ -82,9 +95,13 @@ public int getStatus() {
 }
 
 
+
 public void setStatus(int status) {
 	this.status = status;
 }
+
+
+
